issue 19: Command aggregation working!
