+++ scripts/robo/BallastRunner.php
@@ -1,12 +1,16 @@
 <?php
+
+/**
+ * @file
+ * This functions as a front controller for the Robo commands.
+ */
+
+// @codingStandardsIgnoreFile
+
 $root_directory = dirname(__FILE__, 3);
 require_once $root_directory . '/vendor/autoload.php';
-
-use Consolidation\AnnotatedCommand\CommandFileDiscovery;
-
-$discovery = new CommandFileDiscovery();
-$discovery->setSearchPattern('*.php');
-$commandClasses = $discovery->discover('CommandSupport', '\Ballast');
+$discovery = new Consolidation\AnnotatedCommand\CommandFileDiscovery();
+$commandClasses = $discovery->discover('scripts/robo/src/Commands', '\Ballast\Commands');
 $statusCode = \Robo\Robo::run(
   $_SERVER['argv'],
   $commandClasses,
