+++ setup/ahoy/mac.ahoy.yml
@@ -2,7 +2,7 @@ ahoyapi: v2
 commands:
   robo:
     usage: Call our robo commands [Advanced users only]
-    cmd: vendor/bin/robo --load-from ./scripts/robo "$@"
+    cmd: composer robo "$@"
     hide: true
   bash:
     usage: Open an interactive bash shell in a container [Advanced users only]
@@ -26,16 +26,16 @@ commands:
   cast-off:
     usage: Launch the global tools needed for local development. Run this command once after you boot your computer.
     cmd: |
-      ahoy robo boot && #make sure boot is successful
-      ahoy robo "boot:dns"
+      composer robo boot && #make sure boot is successful
+      composer robo "boot:dns"
   harbor:
     usage: Build the harbor for your docks.  Run this command once after the first time you `composer install` a dp-docker project.
     cmd: |
-      ahoy robo "setup:docker" &&
-      ahoy robo "setup:dns"
+      composer robo "docker:initialize" &&
+      composer robo "docker:proxy-create"
   launch:
     usage: Launch this project site.
-    cmd: ahoy robo docker:compose
+    cmd: composer robo docker:compose
   dock:
     usage: Stops this project site and 'returns to port.'
     cmd: docker-compose $(docker-machine config dp-docker) stop
@@ -59,7 +59,7 @@ commands:
     cmd: docker-compose $(docker-machine config dp-docker) exec front-end node_modules/.bin/gulp build
   rebuild:
     usage: Sync with a server database and compile front end. Pass an environment argument to use with drush alias [@shortname.env]
-    cmd: ahoy robo rebuild "$@"
+    cmd: composer robo rebuild "$@"
   sql:
     usage: Displays connection info for the database service.
-    cmd: ahoy robo connect:sql
+    cmd: composer robo connect:sql
